+++ libs/SmartMeshSDK/IpMgrConnectorSerial/IpMgrConnectorSerialClib.py
@@ -317,7 +317,7 @@ class IpMgrConnectorSerial(ApiConnector,threading.Thread):
             rawVal = getattr(notif.contents,f)
             if  isinstance(rawVal,int):
                 returnVal[f] = rawVal
-            elif isinstance(rawVal,long):
+            elif isinstance(rawVal,int):
                 returnVal[f] = int(rawVal)
             else:
                 returnVal[f] = [int(b) for b in rawVal]
@@ -430,7 +430,7 @@ class IpMgrConnectorSerial(ApiConnector,threading.Thread):
             rawVal = getattr(c_notif.contents,f)
             if  isinstance(rawVal,int):
                 py_notif[f] = rawVal
-            elif isinstance(rawVal,long):
+            elif isinstance(rawVal,int):
                 py_notif[f] = int(rawVal)
             else:
                 py_notif[f] = [int(b) for b in rawVal]
